it wasnt quick and its not fixed
the ball sometimes when shot with max power randomly bounces of nothing. it isnt really bouncing i think its clipping throught the floor and recalcualting and that makes it go up, but when i uped both iteration sovler counters almost nothing changed. the problem is not in wall detector althou it is colliding with the floor but on slowed recording it is clear that it colides with the floor on the landing so nieghter it or the bounciness cause the jump of the ball. tested it all. maybe need to try scaling all the assets and the ball bcs myb the smallness of it all is cooking it. if tomorrow that doesnt work it is becoming a quircky little feature.